+++ modules/elementary_functions/src/cpp/Addition.cpp
@@ -34,6 +34,7 @@
 #include "AdditionComplex.hpp"
 #include "AdditionInteger.hpp"
 #include "FindCommonClass.hpp"
+#include "BinaryOperatorsHelpers.hpp"
 //=============================================================================
 namespace Nelson {
 //=============================================================================
@@ -49,12 +50,7 @@ Addition(const ArrayOf& A, const ArrayOf& B, Class commonClass)
                 res = A;
             }
         } else {
-            Dimensions dimsA = A.getDimensions();
-            Dimensions dimsB = B.getDimensions();
-            if (!(SameSizeCheck(dimsA, dimsB))) {
-                Error(_W("Size mismatch on arguments to arithmetic operator ") + L"+");
-            }
-            res = B;
+            res = binaryOperatorEmptyMatrixEmptryMatrix(A, B, commonClass, "+");
         }
         if (res.getDataClass() == NLS_DCOMPLEX) {
             if (res.allReal()) {